# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Skybull Snapser CloudeCode",
    "description": "Skybull snapser cloud code",
    "contact": {},
    "version": "0.0.1"
  },
...
  "servers": [
    {
      "url": "//localhost:8080/"
    }
  ],
  "paths": {
...
    "/v1/byosnap-skybull-cloudcode/level-up-equipment": {
      "post": {
        "summary": "Level up equipment",
        "description": "Level up equipment",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "sessionToken",
                  "userEquipmentUID",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token"
                  },
                  "userEquipmentUID": {
                    "type": "string",
                    "description": "id of equipment player want to level up"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/byosnap-skybull-cloudcode/re-roll-equipment-modifier": {
      "post": {
        "summary": "Re-Roll equipment modifier",
        "description": "Re-roll player equipment modifier",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "UserEquipmentUID",
                  "sessionToken",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token"
                  },
                  "UserEquipmentUID": {
                    "type": "string",
                    "description": "id of equipment player want to re-roll"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/byosnap-skybull-cloudcode/sell-equipment": {
      "post": {
        "summary": "Sell equipment",
        "description": "Sell player equipment",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "UserEquipmentUID",
                  "sessionToken",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token"
                  },
                  "UserEquipmentUID": {
                    "type": "string",
                    "description": "id of equipment player want to sell"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}